Add lock icon for restricted layers
This adds a lock icon for layers that cannot be previewed or downloaded
either because it is a restricted layer at another institution, or it
is a restricted MIT layer and the user is not logged in. This also
changes the kendallite_user cookie to be an empty string when the user
is anonymous.