Move throttling behavior to search function
This gets rid of the throttle on the function that sets the location
hash and adds a throttle on the search function. In cases where the
location hash is updated in quick succession it was possible for some
updates to be skipped. Throttling the search function should ensure
the location hash is always updated while still rate limiting the Solr
request.